+++ eo-maven-plugin/src/test/java/org/eolang/maven/SkipTest.java
@@ -27,6 +27,7 @@ package org.eolang.maven;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import org.eolang.maven.objectionary.OyFake;
 import org.eolang.maven.util.Home;
 import org.eolang.maven.util.Online;
 import org.hamcrest.MatcherAssert;
