+++ eo-maven-plugin/src/test/java/org/eolang/maven/objectionary/OyFake.java
@@ -21,13 +21,12 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.eolang.maven;
+package org.eolang.maven.objectionary;
 
 import com.jcabi.log.Logger;
 import org.cactoos.Func;
 import org.cactoos.Input;
 import org.cactoos.io.InputOf;
-import org.eolang.maven.objectionary.Objectionary;
 
 /**
  * Objectionary with lambda-function Ctor-s for testing.
