+++ eo-maven-plugin/src/test/java/org/eolang/maven/objectionary/OyFallbackSwapTest.java
@@ -27,7 +27,6 @@ import java.io.IOException;
 import java.nio.file.Path;
 import org.cactoos.io.InputOf;
 import org.cactoos.text.TextOf;
-import org.eolang.maven.OyFake;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.Test;
